pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "m3u8-transcript"
version = "1.3.0"
description = "Download audio from m3u8 streams, transcribe with Whisper, and generate PDF/SRT/TXT transcripts."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Pierre Guirguis"},
]
keywords = ["m3u8", "transcript", "whisper", "pdf", "srt"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Text Processing",
]

dependencies = [
    "openai-whisper>=20231117",
    "yt-dlp>=2024.1.0",
    "ffmpeg-python>=0.2.0",
    "fpdf2>=2.7.0",
    "customtkinter>=5.2.0",
    "packaging>=23.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
m3u8-transcript = "main:main"

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
